+++ base/src/main/java/io/github/sds100/keymapper/base/constraints/ConstraintCategory.kt
@@ -12,5 +12,6 @@ enum class ConstraintCategory {
     PHONE,
     POWER,
     DEVICE,
+    NOTIFICATIONS,
     TIME,
 }