+++ base/src/main/java/io/github/sds100/keymapper/base/constraints/ConstraintId.kt
@@ -54,6 +54,7 @@ enum class ConstraintId {
     HINGE_OPEN,
 
     NOTIFICATION_PANEL_SHOWING,
+    NOTIFICATION_POSTED,
 
     TIME,
 }